Fix:#23 gives appropriate factory pattern to project
