+++ templates/login.html
@@ -16,16 +16,18 @@
                             {% endfor %}
                         {% endif %}
                     {% endwith %}
-                    <form method="POST" action="{{ url_for('login') }}">
-                        <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
+                    <form method="POST" action="{{ url_for('main.login') }}">
+                        {{ form.hidden_tag() }}
+
                         <div class="form-group mb-3">
                             <label for="username">Username</label>
-                            <input type="text" class="form-control" id="username" name="username" required>
+                            {{ form.username(class="form-control", id="username", placeholder="Enter your username") }}
                         </div>
+
                         <div class="form-group mb-3">
                             <label for="password">Password</label>
                             <div class="input-group">
-                                <input type="password" class="form-control" id="password" name="password" required>
+                                {{ form.password(class="form-control", id="password", placeholder="Enter your password") }}
                                 <button class="btn btn-outline-secondary" type="button" tabindex="-1" style="border-left:0; background:transparent;" onclick="togglePassword('password', this)">
                                     <i class="bi bi-eye" id="togglePasswordIcon"></i>
                                 </button>
@@ -33,11 +35,12 @@
                         </div>
 
                         <div class="d-grid gap-2">
-                            <button type="submit" class="btn btn-primary">Login</button>
+                            {{ form.submit(class="btn btn-primary") }}
                         </div>
                     </form>
+
                     <div class="text-center mt-3">
-                        <p>Don't have an account? <a href="{{ url_for('signup') }}">Sign up here</a></p>
+                        <p>Don't have an account? <a href="{{ url_for('main.signup') }}">Sign up here</a></p>
                     </div>
                 </div>
             </div>
